Moved thread handler functions to agent.common and reused in telemetry
