+++ nuvlaedge/agent/agent.py
@@ -19,6 +19,8 @@ from nuvlaedge.agent.job import Job
 from nuvlaedge.agent.orchestrator import COEClient
 from nuvlaedge.agent.orchestrator.factory import get_coe_client
 from nuvlaedge.agent.telemetry import Telemetry
+from nuvlaedge.agent.common.thread_handler import (is_thread_creation_needed,
+                                                   create_start_thread)
 
 
 class Agent:
@@ -263,28 +265,6 @@ class Agent:
             1. Sending heartbeat
             2. Running pull jobs
         """
-
-        def log(level, message, *args, **kwargs):
-            self.logger.log(level, message.format(*args, **kwargs))
-
-        def is_thread_creation_needed(name, thread,
-                                      log_not_exist=(logging.INFO, 'Creating {} thread'),
-                                      log_not_alive=(logging.WARNING, 'Recreating {} thread because it is not alive'),
-                                      log_alive=(logging.DEBUG, 'Thread {} is alive'),
-                                      *args, **kwargs):
-            if not thread:
-                log(*log_not_exist, name, *args, **kwargs)
-            elif not thread.is_alive():
-                log(*log_not_alive, name, *args, **kwargs)
-            else:
-                log(*log_alive, name, *args, **kwargs)
-                return False
-            return True
-
-        def create_start_thread(**kwargs):
-            th = Thread(daemon=True, **kwargs)
-            th.start()
-            return th
 
         if is_thread_creation_needed('Telemetry', self.telemetry_thread,
                                      log_not_alive=(logging.DEBUG, 'Recreating {} thread.'),
